LayoutTests/fast/events/ios/before-input-events-prevent-candidate-insertion-expected.txt


This test verifies that beforeinput events of inputType "insertReplacementText" can be prevented.

To manually test, type 'T' into the text field and try to select a candidate. The replacement text should not be inserted.

Replacing "T" with "The"...
Prevented "insertReplacementText" with data: "The"
Does the input value contain replacement text? false
